+++ src/ensure/index.js
@@ -1,15 +1,11 @@
-import rules from '../runnables';
+import { isRule, globalObject } from '../lib';
+import rules from '../rules';
 import runner from './runner';
-
-const isRule = (rulesObject, name) => (
-    Object.prototype.hasOwnProperty.call(rulesObject, name)
-      && typeof rulesObject[name] === 'function'
-);
 
 function Ensure(customRules = {}) {
     const rulesObject = {...rules, ...customRules};
 
-    if (typeof Proxy === 'function') {
+    if (typeof globalObject.Proxy === 'function') {
         return () => {
 
             const registeredRules = [];
@@ -17,9 +13,11 @@ function Ensure(customRules = {}) {
             const proxy = new Proxy(rulesObject, {
                 get: (rules, ruleName) => {
                     if (ruleName === 'test') {
-                        return (value) => {
-                            return registeredRules.every(({ name, args }) => runner(rules[name], value, ...args));
-                        };
+                        return (value) => (
+                            registeredRules.every(({ name, args }) => (
+                                runner(rules[name], value, ...args))
+                            )
+                        );
                     }
 
                     if (!isRule(rules, ruleName)) { return; }
@@ -37,17 +35,32 @@ function Ensure(customRules = {}) {
         };
     }
 
-    // This is relatively heavier, and preferably should only be done when lacking proxy support
-    return (value) => Object.keys(rulesObject).reduce((allRules, ruleName) => {
-        if (!isRule(rulesObject, ruleName)) { return allRules; }
+    const rulesList = Object.keys(rulesObject);
 
-        allRules[ruleName] = (...args) => {
-            runner(rulesObject[ruleName], value, ...args);
+    return () => {
+        const registeredRules = [];
+
+        return rulesList.reduce((allRules, ruleName) => {
+            if (!isRule(rulesObject, ruleName)) { return; }
+
+            allRules[ruleName] = (...args) => {
+                registeredRules.push({
+                    name: ruleName,
+                    args
+                });
+                return allRules;
+            };
+
             return allRules;
-        };
 
-        return allRules;
-    }, {});
+        }, {
+            test: (value) => (
+                registeredRules.every(({ name, args }) => (
+                    runner(rules[name], value, ...args)
+                ))
+            )
+        });
+    };
 }
 
 const ensure = new Ensure();
