+++ src/lib/globalObject/index.js
@@ -1,0 +1,1 @@
+export default Function('return this')();
