+++ src/lib/index.js
@@ -1,0 +1,3 @@
+export { default as extendRules } from './extendRules';
+export { default as isRule } from './isRule';
+export { default as globalObject } from './globalObject';
