+++ src/lib/isRule/index.js
@@ -1,0 +1,6 @@
+const isRule = (rulesObject, name) => (
+    Object.prototype.hasOwnProperty.call(rulesObject, name)
+      && typeof rulesObject[name] === 'function'
+);
+
+export default isRule;
