+++ src/rules/equals/index.js
