+++ src/rules/equals/spec.js
