+++ src/rules/greaterThan/spec.js
