+++ src/rules/greaterThanOrEquals/index.js
