+++ src/rules/greaterThanOrEquals/spec.js
