+++ src/rules/inside/index.js
