+++ src/rules/inside/spec.js
