+++ src/rules/isArray/index.js
