+++ src/rules/isArray/spec.js
