+++ src/rules/isEmpty/index.js
