+++ src/rules/isEmpty/spec.js
