+++ src/rules/isEven/index.js
