+++ src/rules/isEven/spec.js
