+++ src/rules/isNumber/index.js
