+++ src/rules/isNumber/spec.js
