+++ src/rules/isNumeric/index.js
