+++ src/rules/isNumeric/spec.js
