+++ src/rules/isOdd/index.js
