+++ src/rules/isOdd/spec.js
