+++ src/rules/isString/index.js
