+++ src/rules/isString/spec.js
