+++ src/rules/isTruthy/index.js
