+++ src/rules/isTruthy/spec.js
