+++ src/rules/lengthEquals/index.js
