+++ src/rules/lengthEquals/spec.js
