+++ src/rules/lessThan/index.js
