+++ src/rules/lessThan/spec.js
