+++ src/rules/matches/spec.js
