+++ src/rules/numberEquals/index.js
