+++ src/rules/numberEquals/spec.js
