+++ src/rules/shorterThan/spec.js
