+++ src/rules/spec.js
